pyproject.toml
[project]
name = "llm-playground"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ds-wook", email = "leewook94@gmail.com" }
]
dependencies = [
    "transformers==4.44.0",
    "torch==1.13.1+cu116",
    "accelerate==0.33.0",
    "jinja2>=3.1.3",
    "flake8>=7.1.1",
    "black>=24.10.0",
    "fastapi>=0.115.0",
    "isort>=5.13.2",
    "uvicorn>=0.31.1",
    "nvitop>=1.3.2",
    "google-generativeai>=0.8.3",
    "ipython>=8.28.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.12",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/llm_playground"]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 99

[tool.black]
line-length = 99
target-version = ['py36', 'py37', 'py38']
exclude = '''
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | ^.*\b(migrations)\b.*$
'''
[[tool.rye.sources]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu116"
